# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SkyHigh Core - Digital Check-In System
  description: |
    Backend API for airline self-check-in system.
    
    ## Features
    - Seat selection with conflict-free booking
    - Time-bound seat holds (120 seconds)
    - Baggage validation with payment flow
    - High-performance seat map access
    
    ## Rate Limits
    - Seat map: 100 requests/minute
    - Seat operations: 10 requests/minute
    - Check-in operations: 20 requests/minute
  version: 1.0.0
  contact:
    name: SkyHigh Airlines
    email: api-support@skyhigh.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.skyhigh.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Flights
    description: Flight information
  - name: Seats
    description: Seat management and booking
  - name: Check-in
    description: Check-in workflow
  - name: Baggage
    description: Baggage management
  - name: Payments
    description: Payment processing
  - name: Admin
    description: Administrative operations

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the service is healthy
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/flights:
    get:
      tags:
        - Flights
      summary: List flights
      description: Get a list of flights available for check-in
      operationId: listFlights
      parameters:
        - name: departure_date
          in: query
          description: Filter by departure date
          schema:
            type: string
            format: date
        - name: departure_airport
          in: query
          description: Filter by departure airport code
          schema:
            type: string
            maxLength: 3
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of flights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightListResponse'

  /api/v1/flights/{flight_id}:
    get:
      tags:
        - Flights
      summary: Get flight details
      description: Get detailed information about a specific flight
      operationId: getFlight
      parameters:
        - name: flight_id
          in: path
          required: true
          description: Flight UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Flight details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightResponse'
        '404':
          description: Flight not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/flights/{flight_id}/seats:
    get:
      tags:
        - Seats
      summary: Get seat map
      description: |
        Get the seat map for a flight showing all seats and their availability.
        This endpoint is cached for performance.
      operationId: getSeatMap
      parameters:
        - name: flight_id
          in: path
          required: true
          description: Flight UUID
          schema:
            type: string
            format: uuid
        - name: seat_class
          in: query
          description: Filter by seat class
          schema:
            type: string
            enum: [economy, business, first]
      responses:
        '200':
          description: Seat map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatMapResponse'
        '404':
          description: Flight not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/seats/{seat_id}/hold:
    post:
      tags:
        - Seats
      summary: Hold a seat
      description: |
        Temporarily reserve a seat for 120 seconds.
        The seat will automatically become available if not confirmed within the time limit.
      operationId: holdSeat
      parameters:
        - name: seat_id
          in: path
          required: true
          description: Seat UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HoldSeatRequest'
      responses:
        '200':
          description: Seat held successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldSeatResponse'
        '404':
          description: Seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Seat not available (already held or confirmed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/seats/{seat_id}/release:
    post:
      tags:
        - Seats
      summary: Release a held seat
      description: Release a previously held seat, making it available again
      operationId: releaseSeat
      parameters:
        - name: seat_id
          in: path
          required: true
          description: Seat UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseSeatRequest'
      responses:
        '200':
          description: Seat released successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResponse'
        '400':
          description: Cannot release seat (not held or held by different passenger)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/seats/{seat_id}/confirm:
    post:
      tags:
        - Seats
      summary: Confirm a held seat
      description: |
        Permanently confirm a held seat.
        Can only be done by the passenger who holds the seat.
      operationId: confirmSeat
      parameters:
        - name: seat_id
          in: path
          required: true
          description: Seat UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmSeatRequest'
      responses:
        '200':
          description: Seat confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResponse'
        '400':
          description: Cannot confirm seat (not held or held by different passenger)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Hold expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/checkin/start:
    post:
      tags:
        - Check-in
      summary: Start check-in
      description: |
        Start the check-in process for a passenger.
        This will hold the selected seat and create a check-in record.
      operationId: startCheckIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCheckInRequest'
      responses:
        '201':
          description: Check-in started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Seat not available or passenger already checked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/checkin/{checkin_id}:
    get:
      tags:
        - Check-in
      summary: Get check-in status
      description: Get the current status of a check-in
      operationId: getCheckIn
      parameters:
        - name: checkin_id
          in: path
          required: true
          description: Check-in UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Check-in details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '404':
          description: Check-in not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/checkin/{checkin_id}/complete:
    post:
      tags:
        - Check-in
      summary: Complete check-in
      description: |
        Complete the check-in process.
        Requires all baggage fees to be paid if applicable.
      operationId: completeCheckIn
      parameters:
        - name: checkin_id
          in: path
          required: true
          description: Check-in UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Check-in completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '400':
          description: Cannot complete check-in (invalid state)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment required for excess baggage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Check-in not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Seat hold expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/checkin/{checkin_id}/baggage:
    get:
      tags:
        - Baggage
      summary: Get baggage information
      description: Get baggage information for a check-in
      operationId: getBaggage
      parameters:
        - name: checkin_id
          in: path
          required: true
          description: Check-in UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Baggage information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaggageListResponse'
        '404':
          description: Check-in not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Baggage
      summary: Add baggage
      description: |
        Add baggage to a check-in.
        If total weight exceeds 25kg, the check-in status changes to WAITING_FOR_PAYMENT.
      operationId: addBaggage
      parameters:
        - name: checkin_id
          in: path
          required: true
          description: Check-in UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBaggageRequest'
      responses:
        '201':
          description: Baggage added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaggageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Check-in not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/checkin/{checkin_id}/pay:
    post:
      tags:
        - Payments
      summary: Process payment
      description: Process payment for excess baggage fees
      operationId: processPayment
      parameters:
        - name: checkin_id
          in: path
          required: true
          description: Check-in UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid payment or no payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Check-in not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Payment failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/admin/seed:
    post:
      tags:
        - Admin
      summary: Seed demo data
      description: Populate the database with demo flights, seats, and passengers
      operationId: seedData
      responses:
        '200':
          description: Demo data seeded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedResponse'

  /api/v1/passengers:
    post:
      tags:
        - Passengers
      summary: Create passenger
      description: Create a new passenger record
      operationId: createPassenger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePassengerRequest'
      responses:
        '201':
          description: Passenger created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassengerResponse'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        database:
          type: string
          enum: [connected, disconnected]
        redis:
          type: string
          enum: [connected, disconnected]
        timestamp:
          type: string
          format: date-time

    FlightResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        flight_number:
          type: string
          example: "SH101"
        departure_airport:
          type: string
          example: "JFK"
        arrival_airport:
          type: string
          example: "LAX"
        departure_time:
          type: string
          format: date-time
        arrival_time:
          type: string
          format: date-time
        aircraft_type:
          type: string
          example: "Boeing 737"

    FlightListResponse:
      type: object
      properties:
        flights:
          type: array
          items:
            $ref: '#/components/schemas/FlightResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    SeatResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seat_number:
          type: string
          example: "12A"
        seat_class:
          type: string
          enum: [economy, business, first]
        status:
          type: string
          enum: [AVAILABLE, HELD, CONFIRMED]
        held_by:
          type: string
          format: uuid
          nullable: true
        hold_expires_at:
          type: string
          format: date-time
          nullable: true

    SeatMapResponse:
      type: object
      properties:
        flight_id:
          type: string
          format: uuid
        flight_number:
          type: string
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatResponse'
        summary:
          type: object
          properties:
            total:
              type: integer
            available:
              type: integer
            held:
              type: integer
            confirmed:
              type: integer
        cached:
          type: boolean
        cache_timestamp:
          type: string
          format: date-time
          nullable: true

    HoldSeatRequest:
      type: object
      required:
        - passenger_id
      properties:
        passenger_id:
          type: string
          format: uuid

    HoldSeatResponse:
      type: object
      properties:
        seat:
          $ref: '#/components/schemas/SeatResponse'
        hold_expires_at:
          type: string
          format: date-time
        seconds_remaining:
          type: integer

    ReleaseSeatRequest:
      type: object
      required:
        - passenger_id
      properties:
        passenger_id:
          type: string
          format: uuid

    ConfirmSeatRequest:
      type: object
      required:
        - passenger_id
      properties:
        passenger_id:
          type: string
          format: uuid

    StartCheckInRequest:
      type: object
      required:
        - passenger_id
        - flight_id
        - seat_id
      properties:
        passenger_id:
          type: string
          format: uuid
        flight_id:
          type: string
          format: uuid
        seat_id:
          type: string
          format: uuid

    CheckInResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        passenger_id:
          type: string
          format: uuid
        flight_id:
          type: string
          format: uuid
        seat:
          $ref: '#/components/schemas/SeatResponse'
        status:
          type: string
          enum: [IN_PROGRESS, WAITING_FOR_PAYMENT, COMPLETED]
        baggage:
          type: array
          items:
            $ref: '#/components/schemas/BaggageResponse'
        total_baggage_weight:
          type: number
        excess_fee:
          type: number
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    AddBaggageRequest:
      type: object
      required:
        - weight_kg
      properties:
        weight_kg:
          type: number
          minimum: 0.1
          maximum: 50
          example: 23.5

    BaggageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        weight_kg:
          type: number
        excess_fee:
          type: number
        fee_paid:
          type: boolean

    BaggageListResponse:
      type: object
      properties:
        baggage:
          type: array
          items:
            $ref: '#/components/schemas/BaggageResponse'
        total_weight:
          type: number
        total_excess_fee:
          type: number
        max_allowed_weight:
          type: number
          example: 25

    PaymentRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          minimum: 0.01
        payment_method:
          type: string
          default: "card"
          enum: [card, cash, wallet]

    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
        payment_method:
          type: string
        paid_at:
          type: string
          format: date-time
          nullable: true

    CreatePassengerRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - booking_reference
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 20
        booking_reference:
          type: string
          maxLength: 10

    PassengerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        booking_reference:
          type: string

    SeedResponse:
      type: object
      properties:
        message:
          type: string
        flights_created:
          type: integer
        seats_created:
          type: integer
        passengers_created:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "SEAT_NOT_AVAILABLE"
            message:
              type: string
              example: "Seat 12A is currently held by another passenger"
            details:
              type: object
              additionalProperties: true